# Core: cleanup Item classes