--- a/BaseClasses.py
+++ b/BaseClasses.py
@@ -1153,7 +1153,6 @@ class Item:
     player: int
     location: Optional[Location]
 
-
     def __init__(self, name: str, classification: ItemClassification, code: Optional[int], player: int):
         self.world = None
         self.name = name