--- a/worlds/sm/__init__.py
+++ b/worlds/sm/__init__.py
@@ -5,7 +5,7 @@ import copy
 import os
 import threading
 import base64
-from typing import Set, List, TextIO
+from typing import Set, List, TextIO, Optional
 
 from worlds.sm.variaRandomizer.graph.graph_utils import GraphUtils
 
@@ -735,7 +735,8 @@ class SMLocation(Location):
 
 class SMItem(Item):
     game = "Super Metroid"
+    type: Optional[str] = None
 
-    def __init__(self, name, classification, type, code, player: int = None):
+    def __init__(self, name, classification, type: str, code, player: int = None):
         super(SMItem, self).__init__(name, classification, code, player)
         self.type = type