--- a/worlds/smz3/__init__.py
+++ b/worlds/smz3/__init__.py
@@ -3,7 +3,7 @@ import copy
 import os
 import random
 import threading
-from typing import Dict, Set, TextIO, Optional
+from typing import Dict, Set, TextIO
 
 from BaseClasses import Region, Entrance, Location, MultiWorld, Item, ItemClassification, RegionType, CollectionState, \
     Tutorial
@@ -574,9 +574,9 @@ class SMZ3Location(Location):
 class SMZ3Item(Item):
     game = "SMZ3"
     type: ItemType
-    item = None
+    item: Item
 
-    def __init__(self, name, classification, type: ItemType, code, player: int = None, item=None):
+    def __init__(self, name, classification, type: ItemType, code, player: int, item: Item):
         super(SMZ3Item, self).__init__(name, classification, code, player)
         self.type = type
         self.item = item