Moved all non-common members away from BaseClasses.Item, changed how the alttp credit texts work, fixed / cleaned up / type-hinted some worlds

Feedback is welcome