@black-sliver 2022-08-01
Note that we expect a slight performance regression in IDLE because of added inheritance on items, so i included that in my testing.
Here are some real world numbers running `Generate.py --seed 1`:
```
Alttp solo roll time, 100 repeats
Old: mean 0.880s, min 0.862s, max 0.906s
New: mean 0.878s, min 0.856s, max 0.909s

ArchipIDLE 10 slots roll time, 100 repeats
Old: mean 0.212s, min 0.205s, max 0.221s
New: mean 0.212s, min 0.206s, max 0.225s

Alttp solo peak memory, 10 repeats
Old: mean 55.759MB, min 55.720MB, max 55.780MB
New: mean 55.234MB, min 55.180MB, max 55.250MB

ArchipIDLE 1- slots peak memory, 10 repeats
Old: mean 15.228MB, min 15.220MB, max 15.240MB
New: mean 14.806MB, min 14.800MB, max 14.820MB
```